refactor: move assets, layouts, router, styles and App into app directory
