+++ src/app/assets/fonts/PlusJakartaSans-SemiBold.woff2
