+++ src/app/layouts/app-layout/app-layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import { Header } from '../../components/header/header';
+import { Header } from '../../../components/header/header';
 
 export function AppLayout() {
   return (
