+++ src/app/layouts/courses-layout/courses-layout.tsx
