+++ src/app/router/router.tsx
@@ -1,25 +1,25 @@
 import { createBrowserRouter, Navigate } from 'react-router-dom';
-import { Paths } from './paths';
-import { ErrorPage } from '../pages/error/error-page';
+import { ErrorPage } from '../../pages/error/error-page';
 import { AppLayout } from '../layouts/app-layout/app-layout';
-import { AboutUsPage } from '../pages/about-us/about-us-page';
+import { Paths } from './paths';
+import { AboutUsPage } from '../../pages/about-us/about-us-page';
 import { CoursesLayout } from '../layouts/courses-layout/courses-layout';
-import CoursesPage from '../pages/courses/courses-page';
+import CoursesPage from '../../pages/courses/courses-page';
 import {
   allCoursesLoader,
   courseInfoLoader,
   myCoursesLoader,
   topicLoader,
   topicsLoader,
-} from '../api/loaders';
-import CourseInfoPage from '../pages/course-info/course-info-page';
-import TopicsPage from '../pages/topics/topics-page';
-import TopicPage from '../pages/topic/topic-page';
+} from '../../api/loaders';
+import CourseInfoPage from '../../pages/course-info/course-info-page';
+import TopicsPage from '../../pages/topics/topics-page';
+import TopicPage from '../../pages/topic/topic-page';
 import { WidgetLayout } from '../layouts/widget-layout/widget-layout';
-import { ProfilePage } from '../pages/profile/profile-page';
-import { LoginPage } from '../pages/auth/login/login-page';
-import { RegisterPage } from '../pages/auth/register/register-page';
-import { NotFoundPage } from '../pages/not-found/not-found-page';
+import { ProfilePage } from '../../pages/profile/profile-page';
+import { LoginPage } from '../../pages/auth/login/login-page';
+import { RegisterPage } from '../../pages/auth/register/register-page';
+import { NotFoundPage } from '../../pages/not-found/not-found-page';
 
 export const router = createBrowserRouter([
   {
